+++ mode/node/src/test/java/org/apache/shardingsphere/mode/node/path/type/global/execution/ProcessNodePathTest.java
@@ -27,7 +27,7 @@ class ProcessNodePathTest {
     
     @Test
     void assertToPath() {
-        assertThat(NodePathGenerator.toPath(new ProcessNodePath("foo_process_id", null), false), is("/execution_nodes/foo_process_id"));
-        assertThat(NodePathGenerator.toPath(new ProcessNodePath("foo_process_id", "foo_instance_id"), false), is("/execution_nodes/foo_process_id/foo_instance_id"));
+        assertThat(NodePathGenerator.toPath(new ProcessNodePath("foo_process_id", null)), is("/execution_nodes/foo_process_id"));
+        assertThat(NodePathGenerator.toPath(new ProcessNodePath("foo_process_id", "foo_instance_id")), is("/execution_nodes/foo_process_id/foo_instance_id"));
     }
 }