+++ mode/type/cluster/core/src/main/java/org/apache/shardingsphere/mode/manager/cluster/persist/coordinator/process/ClusterProcessPersistCoordinator.java
@@ -50,9 +50,9 @@ public final class ClusterProcessPersistCoordinator {
     public void reportLocalProcesses(final String instanceId, final String taskId) {
         Collection<Process> processes = ProcessRegistry.getInstance().listAll();
         if (!processes.isEmpty()) {
-            repository.persist(NodePathGenerator.toPath(new ProcessNodePath(taskId, instanceId), false), YamlEngine.marshal(swapper.swapToYamlConfiguration(processes)));
+            repository.persist(NodePathGenerator.toPath(new ProcessNodePath(taskId, instanceId)), YamlEngine.marshal(swapper.swapToYamlConfiguration(processes)));
         }
-        repository.delete(NodePathGenerator.toPath(new ShowProcessListTriggerNodePath(new InstanceProcessNodeValue(instanceId, taskId)), false));
+        repository.delete(NodePathGenerator.toPath(new ShowProcessListTriggerNodePath(new InstanceProcessNodeValue(instanceId, taskId))));
     }
     
     /**
@@ -62,6 +62,6 @@ public final class ClusterProcessPersistCoordinator {
      * @param processId process ID
      */
     public void cleanProcess(final String instanceId, final String processId) {
-        repository.delete(NodePathGenerator.toPath(new KillProcessTriggerNodePath(new InstanceProcessNodeValue(instanceId, processId)), false));
+        repository.delete(NodePathGenerator.toPath(new KillProcessTriggerNodePath(new InstanceProcessNodeValue(instanceId, processId))));
     }
 }