+++ proxy/backend/core/src/test/java/org/apache/shardingsphere/proxy/backend/state/impl/UnavailableProxyStateTest.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.proxy.backend.state.impl;
 
 import org.apache.shardingsphere.distsql.statement.ral.updatable.ImportMetaDataStatement;
-import org.apache.shardingsphere.mode.exception.ClusterStateException;
+import org.apache.shardingsphere.mode.exception.ShardingSphereStateException;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.dml.DMLStatement;
 import org.junit.jupiter.api.Test;
 
@@ -29,7 +29,7 @@ class UnavailableProxyStateTest {
     
     @Test
     void assertExecuteWithUnsupportedSQL() {
-        assertThrows(ClusterStateException.class, () -> new UnavailableProxyState().check(mock(DMLStatement.class)));
+        assertThrows(ShardingSphereStateException.class, () -> new UnavailableProxyState().check(mock(DMLStatement.class)));
     }
     
     @Test